Large Update: - Added Bill DTO for Firebase storing - Edited Icon&Text function to use Double instead of Int - Changed in MainActivity (Rearrangement of elements, bug fixing and new elements) - Payment DTO and Unit Test was restored to their previous state

+++ app/src/main/java/com/rentminder/Icon&Text.kt
@@ -104,7 +104,7 @@ fun OtherIconText() {
 }
 
 @Composable
-fun TotalText(total: Int) {
+fun TotalText(total: Double) {
     Text(
         text = "Total: $total",
         fontSize = (20.sp),
